requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
yt-dlp==2024.01.16
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
requests==2.31.0
tqdm==4.67.0
click==8.1.7
ffmpeg-python==0.2.1
Pillow==10.1.0
alembic==1.13.0
pytz==2024.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
dropbox==11.36.0
spotipy==2.22.1
pytgcalls==3.0.0
python-socks==1.4.0

